+++ lib/screens/home.dart
@@ -5,7 +5,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:sliding_up_panel/sliding_up_panel.dart';
 import 'package:smartflore/bloc/map/map_bloc.dart';
 import 'package:smartflore/bloc/trails/trails_bloc.dart';
-import 'package:smartflore/components/map/panel_widget.dart';
+import 'package:smartflore/components/panel/panel_widget.dart';
 import 'package:smartflore/themes/smart_flore_icons_icons.dart';
 
 import '../components/map/map_widget.dart';
@@ -46,9 +46,7 @@ class _HomeScreenState extends State<HomeScreen> {
                       children: [
                         Container(
                           decoration: const BoxDecoration(
-                              borderRadius:
-                                  BorderRadius.all(Radius.circular(4)),
-                              color: Color(0xFFD8DCD8)),
+                              borderRadius: BorderRadius.all(Radius.circular(4)), color: Color(0xFFD8DCD8)),
                           width: 45,
                           height: 4,
                         ),
@@ -56,32 +54,22 @@ class _HomeScreenState extends State<HomeScreen> {
                         Container(
                           height: 46,
                           decoration: BoxDecoration(
-                              border: Border.all(
-                                  color: Theme.of(context).colorScheme.primary,
-                                  width: 1),
-                              borderRadius:
-                                  const BorderRadius.all(Radius.circular(6))),
+                              border: Border.all(color: Theme.of(context).colorScheme.primary, width: 1),
+                              borderRadius: const BorderRadius.all(Radius.circular(6))),
                           child: TabBar(
                               isScrollable: false,
                               unselectedLabelColor: primary,
                               indicator: BoxDecoration(
-                                  color: primary,
-                                  borderRadius: const BorderRadius.all(
-                                      Radius.circular(6))),
-                              tabs: const [
-                                Tab(text: 'Tous les sentiers'),
-                                Tab(text: 'Mes sentiers')
-                              ]),
+                                  color: primary, borderRadius: const BorderRadius.all(Radius.circular(6))),
+                              tabs: const [Tab(text: 'Tous les sentiers'), Tab(text: 'Mes sentiers')]),
                         ),
                       ],
                     ),
                   ),
                 ),
               ),
-              panelBuilder: (scrollController) =>
-                  _buildSlidingPanel(scrollController: scrollController),
-              borderRadius:
-                  const BorderRadius.vertical(top: Radius.circular(24)),
+              panelBuilder: (scrollController) => _buildSlidingPanel(scrollController: scrollController),
+              borderRadius: const BorderRadius.vertical(top: Radius.circular(24)),
               body: Stack(
                 children: [
                   const MapWidget(),
@@ -115,8 +103,7 @@ class _HomeScreenState extends State<HomeScreen> {
                                     size: 20,
                                   ),
                                   onPressed: () {
-                                    BlocProvider.of<MapBloc>(context)
-                                        .add(RequestCenterMapEvent());
+                                    BlocProvider.of<MapBloc>(context).add(RequestCenterMapEvent());
                                   }),
                             ),
                           ],
@@ -133,8 +120,7 @@ class _HomeScreenState extends State<HomeScreen> {
                   height: 46,
                   child: FloatingActionButton(
                     backgroundColor: Colors.white,
-                    shape: const RoundedRectangleBorder(
-                        borderRadius: BorderRadius.all(Radius.circular(10))),
+                    shape: const RoundedRectangleBorder(borderRadius: BorderRadius.all(Radius.circular(10))),
                     child: const Icon(
                       SmartFloreIcons.setting,
                       size: 20,
